Revert "Merge pull request #88105 from Expensify/revert-87361-perf/extract-money-report-header-actions"
This reverts commit 06e4e74e21f889749d01fe823e601d048cfd28ee, reversing
changes made to 19d70c2a34c978998e3510008478c23a4d8bec9f.
