+++ src/components/MoneyReportHeaderPrimaryAction/types.ts
@@ -5,13 +5,6 @@ type MoneyReportHeaderPrimaryActionProps = {
     reportID: string | undefined;
     chatReportID: string | undefined;
     primaryAction: ValueOf<typeof CONST.REPORT.PRIMARY_ACTIONS> | ValueOf<typeof CONST.REPORT.TRANSACTION_PRIMARY_ACTIONS> | '';
-    isPaidAnimationRunning: boolean;
-    isApprovedAnimationRunning: boolean;
-    isSubmittingAnimationRunning: boolean;
-    stopAnimation: () => void;
-    startAnimation: () => void;
-    startApprovedAnimation: () => void;
-    startSubmittingAnimation: () => void;
     onExportModalOpen: () => void;
 };
