+++ src/hooks/useTransactionThreadReport.ts
@@ -1,0 +1,40 @@
+import getNonEmptyStringOnyxID from '@libs/getNonEmptyStringOnyxID';
+import {getAllNonDeletedTransactions} from '@libs/MoneyRequestReportUtils';
+import {getFilteredReportActionsForReportView, getOneTransactionThreadReportID} from '@libs/ReportActionsUtils';
+import ONYXKEYS from '@src/ONYXKEYS';
+import useNetwork from './useNetwork';
+import useOnyx from './useOnyx';
+import usePaginatedReportActions from './usePaginatedReportActions';
+import useReportTransactionsCollection from './useReportTransactionsCollection';
+
+/**
+ * Derives the single-transaction thread report ID and report for a money request report.
+ *
+ * This pattern is repeated across multiple hooks and components that need to know
+ * whether a report has a single transaction thread (and access its data).
+ */
+function useTransactionThreadReport(reportID: string | undefined) {
+    const {isOffline} = useNetwork();
+
+    const [moneyRequestReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${reportID}`);
+    const [chatReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getNonEmptyStringOnyxID(moneyRequestReport?.chatReportID)}`);
+
+    const {reportActions: unfilteredReportActions} = usePaginatedReportActions(moneyRequestReport?.reportID);
+    const reportActions = getFilteredReportActionsForReportView(unfilteredReportActions);
+
+    const allReportTransactions = useReportTransactionsCollection(reportID);
+    const nonDeletedTransactions = getAllNonDeletedTransactions(allReportTransactions, reportActions, isOffline, true);
+    const visibleTransactions = nonDeletedTransactions?.filter((t) => isOffline || t.pendingAction !== 'delete');
+    const reportTransactionIDs = visibleTransactions?.map((t) => t.transactionID);
+
+    const transactionThreadReportID = getOneTransactionThreadReportID(moneyRequestReport, chatReport, reportActions ?? [], isOffline, reportTransactionIDs);
+    const [transactionThreadReport] = useOnyx(`${ONYXKEYS.COLLECTION.REPORT}${getNonEmptyStringOnyxID(transactionThreadReportID)}`);
+
+    return {
+        transactionThreadReportID,
+        transactionThreadReport,
+        reportActions,
+    };
+}
+
+export default useTransactionThreadReport;
